requirements.txt
overrides
spacy
jsonpickle
editdistance
ftfy
pytorch_pretrained_bert
nltk
unidecode
parsimonious
word2number
conllu
tensorboardX
sqlparse
cvxpy
jsonnet
bert_tokenizer
rake_nltk
gensim
tqdm
